create a chunking abstraction
