+++ interfaces/kalosm-language/src/search/postprocessing/mod.rs
@@ -1,0 +1,4 @@
+// Context extraction strategies:
+// 1. Dump all sentences
+// 2. Dump all sentences that mention an entity
+// 3. Extract relevant sentences with an llm
