+++ interfaces/kalosm/Cargo.toml
@@ -24,6 +24,7 @@ hdrhistogram = "7.5.4"
 num-traits = "0.2.17"
 once_cell = "1.19.0"
 comfy-table = "7.1.0"
+serde = { version = "1.0.163", features = ["derive"] }
 
 [dev-dependencies]
 axum = "0.7.2"
