+++ interfaces/kalosm/examples/fs_context.rs
@@ -8,14 +8,8 @@ async fn main() {
 
     let documents = DocumentFolder::try_from(PathBuf::from("./documents")).unwrap();
 
-    let mut database = DocumentDatabase::new(
-        Bert::builder().build().unwrap(),
-        ChunkStrategy::Sentence {
-            sentence_count: 1,
-            overlap: 0,
-        },
-    );
-    database.extend(documents.clone()).await.unwrap();
+    let mut database = DocumentDatabase::new(Bert::builder().build().unwrap());
+    database.extend(documents.clone()).for_qa().await.unwrap();
     let mut fuzzy = FuzzySearchIndex::default();
     fuzzy.extend(documents).await.unwrap();
